pyproject.toml
[project]
name = "fedlora-poison"
version = "0.1.0"
description = "Layer-targeted model poisoning in federated LoRA fine-tuning of LLMs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flwr>=1.13",
    "peft>=0.14",
    "transformers>=4.45",
    "torch>=2.4",
    "datasets>=3.0",
    "lm-eval>=0.4",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "numpy>=1.26",
    "pandas>=2.1",
    "scikit-learn>=1.4",
    "accelerate>=1.0",
    "safetensors>=0.4",
    "wandb>=0.18",
    "pyyaml>=6.0",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "ipython>=8.0",
    "jupyter>=1.0",
]

[project.scripts]
fedlora = "fedlora_poison.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fedlora_poison"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
